requirements.txt
absl-py==1.3.0
alembic==1.8.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
attrdict==2.0.1
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bob==11.3.0
bob.extension==7.0.3
bob.measure==6.0.2
bokeh==3.0.3
cachetools==5.2.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cloudpickle==2.2.0
cmake==3.24.1.1
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
contourpy==1.0.5
cycler==0.11.0
Cython==0.29.32
databricks-cli==0.17.3
debugpy @ file:///C:/ci_310/debugpy_1642079916595/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dlib @ file:///C:/Users/kimwa/Downloads/Install-dlib-main/Install-dlib-main/dlib-19.22.99-cp310-cp310-win_amd64.whl
docker==6.0.1
easydict==1.10
entrypoints @ file:///C:/ci/entrypoints_1649926676279/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
face-alignment==1.3.5
fire==0.4.0
Flask==2.2.2
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.50.0
h5py==3.7.0
idna==3.4
imageio==2.22.2
importlib-metadata==5.1.0
imutils==0.5.4
ipykernel @ file:///C:/b/abs_21ykzkm7y_/croots/recipe/ipykernel_1662361803478/work
ipython @ file:///C:/ci/ipython_1657652608590/work
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644315428305/work
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
jupyter_client @ file:///C:/b/abs_8fbm7986b_/croots/recipe/jupyter_client_1662504374117/work
jupyter_core @ file:///C:/b/abs_a9330r1z_i/croots/recipe/jupyter_core_1664917313457/work
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
livelossplot==0.5.5
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661934094726/work
mlflow==2.0.1
mtcnn==0.1.1
nest-asyncio @ file:///C:/ci/nest-asyncio_1649848132512/work
networkx==2.8.7
numba==0.56.2
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.5.0
params==0.9.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.6
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyJWT==2.6.0
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorchtools==0.0.2
pytz==2022.4
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1657616000714/work
querystring-parser==1.2.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.1
shap==0.41.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
slicer==0.0.7
smmap==5.0.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
Theano==1.0.5
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchnet==0.0.4
torchsample==0.1.0
torchsummary==1.5.1
torchvision==0.13.1+cu113
tornado @ file:///C:/ci/tornado_1662476985533/work
tqdm==4.64.1
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.4.0
urllib3==1.26.12
visdom==0.2.2
waitress==2.1.2
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websocket-client==1.4.1
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
xgboost==1.7.2
xyzservices==2022.9.0
zipp==3.11.0
